pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email2qa"
version = "0.1.0"
description = "Extract Q&A pairs from Exchange Sent Items"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "exchangelib>=5.4.0",
  "pydantic>=2.9.0",
  "python-dateutil>=2.9.0.post0",
  "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[project.scripts]
email2qa = "email2qa.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
